Web Components Practice
Since I have an idea of a gallery, I am going to use web components to
create a template for a each showing project, basically being similar to
Pinterest Page. I want to practice using web components because I think
it's important to get familiar with that technique. The final mostly
focus on the featured project section